Calculate Net Spread for tows missing net width using a glm.

The Marport Deep Sea Technologies Inc. net mensuration system was used during the deployment of each tow to record net spread and net height. Net width was measured as the horizontal distance between two sensors attached immediately forward of the junction of the upper breastline and the dandyline, and net height was measured from the headrope center to the seafloor. A custom-made AFSC bottom contact sensor (accelerometer) attached to the center of the footrope was used to determine tow duration based on footrope contact with the seafloor. Mean calc_net_spread values for estimating area swept for the tow duration were calculated according to the methods described by Lauth and Kotwicki (2014).

calc_net_spread(dat)

Arguments

dat

data.frame. This can be data from GIDES or race_data.hauls, including these columns: WIRE_OUT, NET_HEIGHT, NET_SPREAD

Value

a list of equations

Details

In race_data, this will manifest as... net_mensuration_code = Net Mensuration Method 0* Unidentified method. Will make racebase.haul$net_mesured = "N". 1 Scanmar net mensuration - don't use, historical 2 NetMind net mensuration - don't use, historical 3 Furuno net mensuration - don't use, historical 4* Estimated from other hauls - when missing net spread, or spread and hieght. Will make racebase.haul$net_mesured = "N". Estimated using GLM 5* Estimated from warp angle - hopefully, will not come up and shouldn't be used 6 Marport net mensuration - shouldn't be there, indicates raw data 7* Marport with sequential outlier rejection, smoothed mean, and adjusted for MKII offset (see AFSC Proc. Report Lauth & Kotwicki 2014). Will make racebase.haul$net_mesured = "Y".

Examples

# Here is an example using 202101 Alaska Night Data from race_data.hauls:
path <- system.file("exdata/calc_net_spread/VESTY202101_RACE_DATA.HAULS.csv",
                    package = "GAPsurvey")
dat <- read.csv(file=path, header=TRUE, sep=",", stringsAsFactors = FALSE)
dat <- dat[dat$PERFORMANCE >= 0 & dat$HAUL_TYPE == 3,]
dat[dat$NET_SPREAD_METHOD != 7, "NET_SPREAD"] <- NA # a normal net width
dat[dat$NET_HEIGHT_METHOD != 6, "NET_HEIGHT"] <- NA # a normal net height
calc_net_spread(dat)
#> $actions
#>          method_col method_code   n
#> 1 NET_SPREAD_METHOD           7 161
#> 2 NET_HEIGHT_METHOD           4   5
#> 3 NET_SPREAD_METHOD           4  27
#>                                                                                                                                                                                                                                     desc
#> 1 There were 161 hauls successfully completed where the Marport sensor properly caclulated net mensuration using sequential outlier rejection, smoothed mean, and MKII offset adjustments (see AFSC Proc. Report Lauth & Kotwicki 2014).
#> 2                                                                                                                     There were 5 missing net height values estimated by averaging the net height of tows with the same wire out scope.
#> 3                                        w ~ 20.835 - 559.975/s - 0.729 * h: For 27 hauls, the net width was estimated using a generalized linear model. The both predictor variables and their interaction were significant (P < 0.001)
#> 
#> $glm_summary
#>                     est0         se     tvalue          prob         var
#> (Intercept)   20.8345889  0.3579184  58.210439 1.179885e-108 (Intercept)
#> INVSCOPE    -559.9746315 38.3797512 -14.590366  4.416356e-31    INVSCOPE
#> NET_HEIGHT    -0.7289979  0.1812255  -4.022601  8.899050e-05  NET_HEIGHT
#>                  est                        case           var0
#> (Intercept)   20.835 very signifcant (P < 0.001)      intercept
#> INVSCOPE    -559.975 very signifcant (P < 0.001) inversed scope
#> NET_HEIGHT    -0.729 very signifcant (P < 0.001)     net height
#> 
#> $dat
#>     HAUL NET_SPREAD NET_HEIGHT WIRE_OUT NET_SPREAD_METHODS NET_HEIGHT_METHOD
#> 2      3   14.83635   2.627722   137.16                  4                 4
#> 3    186   17.86335   2.209000   411.48                  4                 6
#> 4    185   17.92678   2.122000   411.48                  4                 6
#> 5     22   17.45922   2.230000   320.04                  4                 6
#> 6     39   17.19329   2.194773   274.32                  4                 4
#> 7     40   17.18947   2.200000   274.32                  4                 6
#> 8     62   15.00017   2.403000   137.16                  4                 6
#> 9     71   17.91849   1.600000   320.04                  4                 6
#> 10    73   18.04389   1.728000   365.76                  4                 6
#> 11    74   17.95641   1.848000   365.76                  4                 6
#> 12    75   17.97999   2.049000   411.48                  4                 6
#> 13    76   18.25751   1.855000   457.20                  4                 6
#> 14    78   18.22639   1.711000   411.48                  4                 6
#> 15    93   14.83635   2.627722   137.16                  4                 4
#> 16    96   14.83635   2.627722   137.16                  4                 4
#> 17   124   17.07211   2.361000   274.32                  4                 6
#> 18   125   16.52242   2.555000   228.60                  4                 6
#> 19   127   17.16177   2.238000   274.32                  4                 6
#> 20   130   16.54429   2.525000   228.60                  4                 6
#> 21   132   16.52752   2.548000   228.60                  4                 6
#> 22   155   17.75375   2.126000   365.76                  4                 6
#> 23   157   17.65096   2.267000   365.76                  4                 6
#> 24   158   17.80186   2.060000   365.76                  4                 6
#> 25   162   17.70928   2.187000   365.76                  4                 6
#> 26   175   17.38997   2.325000   320.04                  4                 6
#> 27   179   17.80649   2.287000   411.48                  4                 6
#> 28   180   17.82472   2.262000   411.48                  4                 6
#> 39   153   17.17460   2.343000   365.76                  7                 6
#> 40   154   18.06890   2.172000   411.48                  7                 6
#> 41   156   17.26440   2.022000   365.76                  7                 6
#> 42   160   17.65670   2.246000   320.04                  7                 6
#> 43   161   17.72740   2.305000   365.76                  7                 6
#> 44   163   16.91930   2.240000   365.76                  7                 6
#> 45   164   17.80950   2.166000   365.76                  7                 6
#> 46   165   17.85030   2.193000   320.04                  7                 6
#> 47   166   17.89860   2.234000   320.04                  7                 6
#> 48   167   17.89930   2.080000   365.76                  7                 6
#> 49   168   17.82630   2.109000   365.76                  7                 6
#> 50   169   18.03190   2.031000   365.76                  7                 6
#> 51   170   17.90640   2.016000   320.04                  7                 6
#> 52   171   17.30380   2.134000   320.04                  7                 6
#> 53   172   17.63530   2.083000   320.04                  7                 6
#> 54   173   18.04400   2.273000   320.04                  7                 6
#> 55   174   17.73700   2.184000   320.04                  7                 6
#> 56   176   17.63460   2.077000   365.76                  7                 6
#> 57   177   18.20400   2.234000   365.76                  7                 6
#> 58   178   18.22860   2.259000   365.76                  7                 6
#> 59   181   19.17180   2.186000   411.48                  7                 6
#> 60   182   17.95510   1.952000   411.48                  7                 6
#> 61   183   18.34570   2.080000   411.48                  7                 6
#> 62   184   17.71440   2.077000   365.76                  7                 6
#> 63   187   17.69960   2.038000   365.76                  7                 6
#> 64   188   17.84100   2.052000   365.76                  7                 6
#> 65   189   18.36430   2.225000   320.04                  7                 6
#> 66   190   17.89060   2.090000   274.32                  7                 6
#> 67   191   18.83180   1.968000   274.32                  7                 6
#> 68   192   17.46850   2.266000   274.32                  7                 6
#> 69   193   17.28910   2.285000   228.60                  7                 6
#> 70   194   17.46780   2.278000   228.60                  7                 6
#> 71   195   17.40940   2.351000   228.60                  7                 6
#> 103   79   17.10380   1.842000   411.48                  7                 6
#> 104   80   17.27980   1.827000   320.04                  7                 6
#> 105   81   16.65240   1.903000   274.32                  7                 6
#> 106   82   15.83280   2.174000   228.60                  7                 6
#> 107   83   15.52930   2.013000   228.60                  7                 6
#> 108   84   16.48600   1.860000   228.60                  7                 6
#> 109   85   16.31050   1.917000   228.60                  7                 6
#> 110   86   15.99340   1.888000   182.88                  7                 6
#> 111   87   15.80450   2.250000   228.60                  7                 6
#> 112   88   15.71270   2.299000   182.88                  7                 6
#> 113   89   15.62770   2.401000   182.88                  7                 6
#> 114   90   15.07380   2.254000   137.16                  7                 6
#> 115   91   15.75430   2.166000   137.16                  7                 6
#> 116   92   15.45820   2.171000   137.16                  7                 6
#> 117   94   15.28100   2.148000   137.16                  7                 6
#> 118   95   15.35290   2.105000   137.16                  7                 6
#> 119   97   15.37120   2.454000   137.16                  7                 6
#> 120   98   15.82680   2.563000   137.16                  7                 6
#> 121   99   15.76280   1.911000   137.16                  7                 6
#> 122  100   16.43660   2.198000   182.88                  7                 6
#> 123  101   16.71100   2.107000   182.88                  7                 6
#> 124  102   14.94470   2.025000   182.88                  7                 6
#> 125  103   16.60100   2.058000   182.88                  7                 6
#> 126  104   17.11460   1.820000   228.60                  7                 6
#> 127  105   16.61830   2.025000   228.60                  7                 6
#> 128  106   16.68170   1.858000   228.60                  7                 6
#> 129  107   16.39690   1.914000   228.60                  7                 6
#> 130  108   16.23050   2.292256   228.60                  7                 4
#> 131  109   17.24800   2.105000   228.60                  7                 6
#> 132  110   16.16760   2.041000   228.60                  7                 6
#> 133  111   15.97460   2.093000   228.60                  7                 6
#> 134  112   15.52170   2.191000   228.60                  7                 6
#> 135  113   16.57390   2.099000   274.32                  7                 6
#> 136  114   15.57140   2.313000   228.60                  7                 6
#> 137  115   15.64320   2.316000   274.32                  7                 6
#> 138  117   16.55170   1.943000   274.32                  7                 6
#> 139  118   17.19710   1.905000   320.04                  7                 6
#> 140  119   16.47080   2.087000   228.60                  7                 6
#> 141  120   16.10490   2.134000   228.60                  7                 6
#> 142  121   15.67950   1.986000   182.88                  7                 6
#> 143  122   15.10710   2.223000   182.88                  7                 6
#> 144  123   15.97710   1.977000   228.60                  7                 6
#> 145  126   16.59840   2.554000   274.32                  7                 6
#> 146  128   17.23980   2.637000   274.32                  7                 6
#> 147  129   16.30940   2.701000   228.60                  7                 6
#> 148  131   17.07920   2.561000   228.60                  7                 6
#> 149  133   16.89850   2.652000   228.60                  7                 6
#> 150  134   16.34050   2.543000   228.60                  7                 6
#> 151  135   17.50510   2.864000   228.60                  7                 6
#> 152  136   18.53180   2.660000   228.60                  7                 6
#> 153  137   16.72080   3.029000   228.60                  7                 6
#> 154  138   16.98170   2.740000   228.60                  7                 6
#> 155  139   16.60090   2.868000   228.60                  7                 6
#> 156  140   16.38400   2.904000   228.60                  7                 6
#> 157  142   18.23070   2.173000   274.32                  7                 6
#> 158  143   18.24750   2.116000   274.32                  7                 6
#> 159  144   16.85120   2.248000   274.32                  7                 6
#> 160  145   18.21740   2.235000   274.32                  7                 6
#> 161  146   18.09960   2.209000   274.32                  7                 6
#> 162  147   17.75310   2.054000   274.32                  7                 6
#> 163  148   18.15540   2.120000   320.04                  7                 6
#> 164  149   17.98690   2.092000   320.04                  7                 6
#> 165  150   17.93780   2.021000   320.04                  7                 6
#> 166  151   19.46440   2.002000   320.04                  7                 6
#> 167    4   15.25780   2.467000   182.88                  7                 6
#> 168    5   14.36380   2.934000   137.16                  7                 6
#> 169    6   15.27740   2.729000   182.88                  7                 6
#> 170    7   14.69230   3.084000   137.16                  7                 6
#> 171    8   14.67880   2.706000   137.16                  7                 6
#> 172    9   14.93400   2.718000   137.16                  7                 6
#> 173   10   15.27550   2.681000   137.16                  7                 6
#> 174   11   15.96140   2.429000   182.88                  7                 6
#> 175   12   14.74250   2.874000   137.16                  7                 6
#> 176   13   13.82240   2.893000   137.16                  7                 6
#> 177   14   15.12340   2.695000   182.88                  7                 6
#> 178   15   14.93340   2.822000   137.16                  7                 6
#> 179   16   15.26170   2.688000   182.88                  7                 6
#> 180   17   15.65620   2.489000   182.88                  7                 6
#> 181   19   15.36790   2.572000   182.88                  7                 6
#> 182   20   15.23390   2.694000   182.88                  7                 6
#> 183   21   16.17880   2.467000   228.60                  7                 6
#> 184   23   16.68610   2.427000   274.32                  7                 6
#> 185   24   16.41350   2.445000   228.60                  7                 6
#> 186   25   15.51110   2.530000   182.88                  7                 6
#> 187   26   15.70080   2.379000   182.88                  7                 6
#> 188   27   15.82300   2.555000   137.16                  7                 6
#> 189   28   16.42010   2.439000   182.88                  7                 6
#> 190   29   14.76970   2.709000   137.16                  7                 6
#> 191   30   12.79130   3.414000   137.16                  7                 6
#> 192   31   15.49810   2.530000   137.16                  7                 6
#> 193   32   15.18150   2.869000   137.16                  7                 6
#> 194   33   15.37120   2.792000   137.16                  7                 6
#> 195   34   15.94770   2.386000   137.16                  7                 6
#> 196   35   16.50700   2.606000   182.88                  7                 6
#> 197   36   16.26920   2.395000   182.88                  7                 6
#> 198   37   16.64200   2.406000   228.60                  7                 6
#> 199   38   16.52550   2.492000   228.60                  7                 6
#> 200   41   17.08090   2.333000   320.04                  7                 6
#> 201   42   16.89840   2.316000   320.04                  7                 6
#> 202   43   16.18790   2.308000   320.04                  7                 6
#> 203   44   17.22730   2.374000   320.04                  7                 6
#> 204   45   17.56640   2.306000   274.32                  7                 6
#> 205   46   16.68040   2.226000   274.32                  7                 6
#> 206   47   16.17150   2.412000   228.60                  7                 6
#> 207   48   16.42820   2.348000   228.60                  7                 6
#> 208   49   15.92490   2.410000   182.88                  7                 6
#> 209   50   15.71800   2.559000   182.88                  7                 6
#> 210   51   15.87330   2.437000   182.88                  7                 6
#> 211   52   15.10380   2.709000   137.16                  7                 6
#> 212   53   14.66880   2.792000   137.16                  7                 6
#> 213   54   15.05630   2.662000   137.16                  7                 6
#> 214   55   13.95290   3.082000   137.16                  7                 6
#> 215   56   14.57480   2.768000   137.16                  7                 6
#> 216   57   14.30510   2.963000   137.16                  7                 6
#> 217   58   12.56630   3.249000   137.16                  7                 6
#> 218   59   14.91530   2.698000   137.16                  7                 6
#> 219   60   15.28120   2.655000   137.16                  7                 6
#> 220   61   15.22370   2.883000   137.16                  7                 6
#> 221   63   15.58820   1.933000   137.16                  7                 6
#> 222   64   15.92360   2.062000   137.16                  7                 6
#> 223   65   16.76920   1.860000   182.88                  7                 6
#> 224   66   17.35500   1.719000   228.60                  7                 6
#> 225   67   17.90080   1.861000   228.60                  7                 6
#> 226   68   17.36020   1.761000   228.60                  7                 6
#> 227   69   17.46470   1.820000   228.60                  7                 6
#> 228   70   16.92950   1.716000   274.32                  7                 6
#> 229   72   17.78080   1.740000   320.04                  7                 6
#> 230   77   18.56800   1.814000   411.48                  7                 6
#>     WIRE_OUT_METHODS
#> 2                  5
#> 3                  5
#> 4                  5
#> 5                  5
#> 6                  5
#> 7                  5
#> 8                  5
#> 9                  5
#> 10                 5
#> 11                 5
#> 12                 5
#> 13                 5
#> 14                 5
#> 15                 5
#> 16                 5
#> 17                 5
#> 18                 5
#> 19                 5
#> 20                 5
#> 21                 5
#> 22                 5
#> 23                 5
#> 24                 5
#> 25                 5
#> 26                 5
#> 27                 5
#> 28                 5
#> 39                 5
#> 40                 5
#> 41                 5
#> 42                 5
#> 43                 5
#> 44                 5
#> 45                 5
#> 46                 5
#> 47                 5
#> 48                 5
#> 49                 5
#> 50                 5
#> 51                 5
#> 52                 5
#> 53                 5
#> 54                 5
#> 55                 5
#> 56                 5
#> 57                 5
#> 58                 5
#> 59                 5
#> 60                 5
#> 61                 5
#> 62                 5
#> 63                 5
#> 64                 5
#> 65                 5
#> 66                 5
#> 67                 5
#> 68                 5
#> 69                 5
#> 70                 5
#> 71                 5
#> 103                5
#> 104                5
#> 105                5
#> 106                5
#> 107                5
#> 108                5
#> 109                5
#> 110                5
#> 111                5
#> 112                5
#> 113                5
#> 114                5
#> 115                5
#> 116                5
#> 117                5
#> 118                5
#> 119                5
#> 120                5
#> 121                5
#> 122                5
#> 123                5
#> 124                5
#> 125                5
#> 126                5
#> 127                5
#> 128                5
#> 129                5
#> 130                5
#> 131                5
#> 132                5
#> 133                5
#> 134                5
#> 135                5
#> 136                5
#> 137                5
#> 138                5
#> 139                5
#> 140                5
#> 141                5
#> 142                5
#> 143                5
#> 144                5
#> 145                5
#> 146                5
#> 147                5
#> 148                5
#> 149                5
#> 150                5
#> 151                5
#> 152                5
#> 153                5
#> 154                5
#> 155                5
#> 156                5
#> 157                5
#> 158                5
#> 159                5
#> 160                5
#> 161                5
#> 162                5
#> 163                5
#> 164                5
#> 165                5
#> 166                5
#> 167                5
#> 168                5
#> 169                5
#> 170                5
#> 171                5
#> 172                5
#> 173                5
#> 174                5
#> 175                5
#> 176                5
#> 177                5
#> 178                5
#> 179                5
#> 180                5
#> 181                5
#> 182                5
#> 183                5
#> 184                5
#> 185                5
#> 186                5
#> 187                5
#> 188                5
#> 189                5
#> 190                5
#> 191                5
#> 192                5
#> 193                5
#> 194                5
#> 195                5
#> 196                5
#> 197                5
#> 198                5
#> 199                5
#> 200                5
#> 201                5
#> 202                5
#> 203                5
#> 204                5
#> 205                5
#> 206                5
#> 207                5
#> 208                5
#> 209                5
#> 210                5
#> 211                5
#> 212                5
#> 213                5
#> 214                5
#> 215                5
#> 216                5
#> 217                5
#> 218                5
#> 219                5
#> 220                5
#> 221                5
#> 222                5
#> 223                5
#> 224                5
#> 225                5
#> 226                5
#> 227                5
#> 228                5
#> 229                5
#> 230                5
#>