Convert CTD .hex file to .cnv using SBE Data Processing

Converts SBE19plus V2 CTD hex file to cnv using Sea-Bird Data Processing software. cnv files generated with this function will be properly formatted for conversion to BTD using convert_ctd_btd().

convert_ctd_hex(
  hex_file_path,
  xmlcon_path,
  bat_file = NULL,
  datcnv_file = NULL
)

Arguments

hex_file_path

Path to a single .hex file from the CTD.

xmlcon_path

Path to CTD the config file (.xmlcon) for the CTD that was used to collect data in the .hex file.

bat_file

Optional. Filepath to a batch (.bat) file to be executed in command line to run SBE Data Processing. Default NULL automatically uses the .bat file that is included in GAPsurvey.

datcnv_file

Optional. Filepath to .psa file for the SBE Data Processing Datcnv module. Default NULL automatically uses the .psa file that is included in GAPsurvey.

Value

Writes decoded CTD data to a cnv file and returns the path to the file as a 1L character vector.

Examples

# Copy system files to working directory for example
file.copy(from = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
package = "GAPsurvey"),
to = gsub(pattern = system.file("exdata/convert_ctd_btd/", package = "GAPsurvey"),
          replacement = getwd(),
          x = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
          package = "GAPsurvey")))
#> [1] FALSE

file.copy(from = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                             package = "GAPsurvey"),
          to = gsub(pattern = system.file("exdata/convert_ctd_btd/",
                                           package = "GAPsurvey"),
                                           replacement = getwd(),
                    x = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                                    package = "GAPsurvey")))
#> [1] FALSE

# Run convert_ctd_hex()
GAPsurvey::convert_ctd_hex(hex_file_path = "2021_06_13_0003.hex",
                            xmlcon_path = "19-8102_Deploy2021.xmlcon",
                            bat_file = NULL,
                            datcnv_file = NULL)
#> convert_ctd_hex: Attempting to convert C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/2021_06_13_0003.hex to .cnv
#> convert_ctd_hex: .cnv file created at C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/2021_06_13_0003_raw.cnv.
#> [1] "C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/2021_06_13_0003_raw.cnv"