Extract depth and temperature from NMEA strings

Convert Marport Trawl Explorer NMEA strings to BTD and BTH files. Called internally by convert_bvdr_marp().

Usage

convert_nmea_btd(
  nmea_strings = NULL,
  filter_type = "none",
  min_depth = -0.1,
  max_depth = 800,
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = "Marport Trawl Explorer",
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  ...
)

Arguments

nmea_strings

Character vector of NMEA strings (e.g., from a .bvdr file) or a path to a .marp file.

filter_type

Should depth and temperature channels use a 5-scan median window filter ("median"), low-pass filter ("lowpass"), or no filter ("none") be applied to temperature and depth data to remove erroneous outliers? Default = TRUE.

min_depth

Optional (default = -0.1). Minimum valid depth value.

max_depth

Optional (default = 1000). Maximum valid depth value.

VESSEL

Optional. Default = NA. The vessel number (e.g., 162 for AK Knight, 94 for Vesteraalen). If NA or not called in the function, a prompt will appear asking for this data.

CRUISE

Optional. Default = NA. The cruise number, which is usually the year + sequential two digit cruise (e.g., 202101). If NA or not called in the function, a prompt will appear asking for this data.

HAUL

Optional. Default = NA. The haul number that you are trying to convert data for (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.

MODEL_NUMBER

Optional. Default = "Marport Trawl Explorer". The model name/number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).

VERSION_NUMBER

Optional. Default = NA. The version number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual version number without any negative repercussions).

SERIAL_NUMBER

Optional. Default = NA. The serial number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual serial number without any negative repercussions).

...

additional arguments

Author

Sean Rohan <sean.rohan@noaa.gov>

Examples

if (FALSE) { # \dontrun{
# Run this to select Marport (.marp files)
convert_nmea_btd()
} # }