BVDR Conversion to Create BTD data

Converts Marport BVDR data (.ted and .tet files from Marport headrope sensor) to .BTD format. You must first run the BVDR converter program (convert_bvdr.exe) to convert the Marport .bvdr files into .ted and .tet files that can be pulled into R. The BVDR program and instructions can be found in the RACE Survey App. You will have to create your own .SGT file using the example in the BVDR instruction file with start and end time (be sure to include a carriage return after your (second and) final row of data!), because this is not a file that our current systems creates. Once you have used the BVDR converter to output the .ted and .tet files you are ready to use the convert_ted_btd() function here!

Usage

convert_ted_btd(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in = "C:/Program Files/Marport Server/Logs/",
  path_out = "./",
  filename_add = "new"
)

Arguments

VESSEL

Optional. Default = NA. The vessel number (e.g., 162 for AK Knight, 94 for Vesteraalen). If NA or not called in the function, a prompt will appear asking for this data.

CRUISE

Optional. Default = NA. The cruise number, which is usually the year + sequential two digit cruise (e.g., 202101). If NA or not called in the function, a prompt will appear asking for this data.

HAUL

Optional. Default = NA. The haul number that you are trying to convert data for (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.

MODEL_NUMBER

Optional. Default = NA. The model number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).

VERSION_NUMBER

Optional. Default = NA. The version number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual version number without any negative repercussions).

SERIAL_NUMBER

Optional. Default = NA. The serial number of the Marport sensor (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual serial number without any negative repercussions).

path_in

Optional. The default is the location on the catch computer ("C:/Program Files/Marport Server/Logs/") but any path can be entered.

path_out

Optional. The default is the local working directory but can be specified with a string.

filename_add

Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.

Value

.BTH and .BTD files to the path_out directory.

Examples

convert_ted_btd(
   VESSEL = 94,
   CRUISE = 201901,
   HAUL = 3,
   MODEL_NUMBER = 123,
   VERSION_NUMBER = 456,
   SERIAL_NUMBER = 789,
   path_in = system.file("exdata/convert_bvdr_btd/", package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newted")
#> Your new Z:/Projects/GAPsurvey_general/GAPsurvey/docs/reference/HAUL0003_newted .BTD and .BTH files are saved.