Import Length Tablet Files Into DataEnt.mdb

NOTE: Must be run on 32-bit R! NOTE: This function is typically run within the catchData()

lengthData(haul, dsnTablet, dsnDataEnt)

Arguments

haul
Number of the haul you want to upload

dsnTablet
String for the drive folder (in R notation) where tablet files exist.

dsnDataEnt
String for the drive folder (in R notation) - include full file name and extension.

Value

changes to catch data in the access database (.mbd)

Details

Looks for tablet length files in given DSN and uploads them to LENGTH and RAW_LENGTH in the given data_ent.mdb.

Examples

# Gulf of Alaska Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 8

# Import length data
lengthData(haul,dsnTablet,dsnDataEnt)  

#> Warning: [RODBC] ERROR: state IM002, code 0, message [Microsoft][ODBC Driver Manager] Data source name not found and no default driver specified

#> Warning: ODBC connection failed

#> Error in RODBC::sqlQuery(dataEnt, query): first argument is not an open RODBC channel

# Eastern Bering Sea Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/EBS/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/EBS/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 59

# Import length data
lengthData(haul,dsnTablet,dsnDataEnt) 

#> Warning: [RODBC] ERROR: state IM002, code 0, message [Microsoft][ODBC Driver Manager] Data source name not found and no default driver specified

#> Warning: ODBC connection failed

#> Error in RODBC::sqlQuery(dataEnt, query): first argument is not an open RODBC channel