Import Specimen Tablet Files Into DataEnt.mdb

Looks for tablet specimen files in given DSN and uploads them to SPECIMEN in the given data_ent.mdb.

specimenData(haul, dsnTablet, dsnDataEnt)

Arguments

haul
Number of the haul you want to upload

dsnTablet
String for the drive folder (in R notation) where tablet files exist.

dsnDataEnt
String for the drive folder (in R notation) - include full file name and extension.

Value

changes to catch data in the access database (.mbd)

Details

#---> To use the program you must first: Bluetooth transfer the following files from tablets to catch computer (should automatically go into C:Users/NOAADATA/Documents/Bluetooth/inbox) Each of these file names will begin with the tablet name and end with the haul number. DATAENT_CATCH_xxxx.csv DATAENT_BBAG_xxxx.csv -- If there was a benthic bag RAW_BBAG_xxxx.csv -- If there was a benthic bag RAW_CATCH_HAUL_xxxx.csv RAW_CATCH_HAUL_ATTR_xxxx.csv RAW_CATCH_SAMPLE_xxxx.csv RAW_CATCH_VALUE_xxxx.csv SPECIMEN_xxxx.csv -- The specimen data HAUL_xxxx.csv -- A length file for each tablet

PLEASE NOTE: You cannot have any species in your catch data files that have a blank subsample_weight! If you erroneously tap on a species during data collection, you must delete that species from the tablet before transfering the data to the catch computer!

Examples

# Gulf of Alaska Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/GOA/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/GOA/data_ent.mdb", package = "GAPsurvey")
importLength <- TRUE ## Change this "T" to "F" if not importing length data ##

# Define the current haul number
haul <- 8

# Import specimen data
specimenData(haul,dsnTablet,dsnDataEnt)

#> [1] 1

# Eastern Bering Sea Example
# Point functions to correct tablet directory and "dataEnt" file
dsnTablet <- system.file("exdata/catch/EBS/", package = "GAPsurvey")
dsnDataEnt <- system.file("exdata/catch/EBS/data_ent.mdb", package = "GAPsurvey")

# Define the current haul number
haul <- 59

# Import specimen data
specimenData(haul,dsnTablet,dsnDataEnt)

#> [1] 1