Convert CTD data in .cnv form to BTD and BTH

Before running this convert_ctd_btd function, you will need to use the CTD laptop to convert the raw CTD data to a .cnv file.

convert_ctd_btd(
  VESSEL = NA,
  CRUISE = NA,
  HAUL = NA,
  MODEL_NUMBER = NA,
  VERSION_NUMBER = NA,
  SERIAL_NUMBER = NA,
  path_in,
  path_out = "./",
  path_xmlcon = NULL,
  filename_add = ""
)

Arguments

VESSEL

Optional. Default = NA. The vessel number (e.g., 94). If NA or not called in the function, a prompt will appear asking for this data.

CRUISE

Optional. Default = NA. The cruise number, which is usually the year date (e.g., 201901). If NA or not called in the function, a prompt will appear asking for this data.

HAUL

Optional. Default = NA. The haul number, aka the iterative number of this haul (e.g., 3). If NA or not called in the function, a prompt will appear asking for this data.

MODEL_NUMBER

Optional. Default = NA. The model number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).

VERSION_NUMBER

Optional. Default = NA. The version number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).

SERIAL_NUMBER

Optional. Default = NA. The serial number of the CTD (e.g., 123 or 999, you can put in NA or a dummy number here instead of the actual model number without any negative repercussions).

path_in

Optional. Default = "./., or the local working directory but any path (as a string) may be entered. Can accept a .hex file if SBE Data Processing software is installed and path_xmlcon is specified.

path_out

Optional. The default is the local working directory but may be specified with a string.

path_xmlcon

Optional. Filepath to the CTD configuration file (.xmlcon)

filename_add

Optional. Default = "new". This string will be added to the name of the outputed file. Here, you can additional information that may make this file helpful to find later.

Value

.BTH and .BTD files to the path_out directory.

Details

To do this, 1. use the SBE Data Processing Program, and in the "Run" menu select "1. Data conversion". 2. Under "Program setup file" select the .psa file (should be DataCnv.psa located in CTD folder), under "Instrument configuration file" select the .xmlcon file for the CTD located in the deployment_xmlcon folder- this is CTD specific, know your CTD's serial number, and under "Input directory" select the .hex file (from the CTD computer) that is specific to the haul that you are missing data. Click "Start Process". The conversion program will ask you for outputs. At the prompt, Select and Add "Time, Elapsed, seconds", "Depth, salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m". 3. This .cnv file must include columns for "Time, Elapsed, seconds", "Depth , salt water, m", "Temperature, ITS-90, deg C", "Pressure, Strain Gauge, db", and "Conductivity, S/m". 4. Press "Start Process" button again and a .cnv file should appear in your selected output directory. The .cnv file can then be used for this convert_ctd_btd() function! 5. Look in your output directory (usually the "Documents" folder, but you can find it using getwd()) for your new .BTD and .BTH files.

Note that if there are multiple observations from the CTD per second, that they will be averaged by second (e.g., any observations from seconds 0 to >1.0 will be averaged together).

Examples

convert_ctd_btd(
   VESSEL = 94,
   CRUISE = 2001,
   HAUL = 4,
   MODEL_NUMBER = 123, # for example, also can make up
   VERSION_NUMBER = 456, # for example, also can make up
   SERIAL_NUMBER = 789, # for example, also can make up
   path_in = system.file(paste0("exdata/convert_ctd_btd/",
      "SBE19plus_01908103_2021_06_01_94_0004_raw.cnv"),
       package = "GAPsurvey"),
   path_xmlcon = system.file(paste0("exdata/convert_ctd_btd/",
     "19-8102_Deploy2021.xmlcon"),
      package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd")
#> Your new C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/HAUL0004_newctd .BTD and .BTH files are saved.

convert_ctd_btd(
   VESSEL = 94,
   CRUISE = 202101,
   HAUL = 107,
   MODEL_NUMBER = "", # for example, also can make up
   VERSION_NUMBER = "", # for example, also can make up
   SERIAL_NUMBER = 8105, # for example, also can make up
   path_in = system.file(paste0("exdata/convert_ctd_btd/",
     "2021_06_13_0003.hex"),
      package = "GAPsurvey"),
   path_xmlcon = system.file(paste0("exdata/convert_ctd_btd/",
     "19-8102_Deploy2021.xmlcon"),
      package = "GAPsurvey"),
   path_out = getwd(),
   filename_add = "newctd")
#> convert_ctd_hex: Attempting to convert C:/Users/emily.markowitz/AppData/Local/Temp/Rtmp824haa/temp_libpath87817f05a05/GAPsurvey/exdata/convert_ctd_btd/2021_06_13_0003.hex to .cnv
#> convert_ctd_hex: .cnv file created at C:/Users/emily.markowitz/AppData/Local/Temp/Rtmp824haa/temp_libpath87817f05a05/GAPsurvey/exdata/convert_ctd_btd/2021_06_13_0003_raw.cnv.
#> Your new C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/HAUL0107_newctd .BTD and .BTH files are saved.

# Copy system files to working directory for example
file.copy(
   from = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
                      package = "GAPsurvey"),
   to = gsub(pattern = system.file("exdata/convert_ctd_btd/",
             package = "GAPsurvey"),
             replacement = getwd(),
             x = system.file("exdata/convert_ctd_btd/2021_06_13_0003.hex",
                                     package = "GAPsurvey")))
#> [1] FALSE

file.copy(
   from = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                       package = "GAPsurvey"),
          to = gsub(pattern = system.file("exdata/convert_ctd_btd/",
                    package = "GAPsurvey"),
                    replacement = getwd(),
                    x = system.file("exdata/convert_ctd_btd/19-8102_Deploy2021.xmlcon",
                                    package = "GAPsurvey")))
#> [1] FALSE

# Run convert_ctd_btd on .hex file and .xmlcon file
GAPsurvey::convert_ctd_btd(path_in = "2021_06_13_0003.hex",
                   path_xmlcon = "19-8102_Deploy2021.xmlcon")
#> Type vessel code:  
#> Type cruise number:  
#> Type haul number:  
#> Type model number:  
#> Type version number:  
#> Type serial number of CTD:  
#> convert_ctd_hex: Attempting to convert C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/2021_06_13_0003.hex to .cnv
#> convert_ctd_hex: .cnv file created at C:/Users/emily.markowitz/Documents/Projects/GAPsurvey_general/GAPsurvey/docs/reference/2021_06_13_0003_raw.cnv.
#> Your new ./HAULNANA .BTD and .BTH files are saved.
# Convert directly from .hex file by using convert_ctd_btd to run SBE Data Processing.