Get sunrise and sunset times by day, latitude, and longitude

Get sunrise and sunset times by day, latitude, and longitude

get_sunrise_sunset(
  chosen_date,
  latitude = NULL,
  longitude = NULL,
  survey = NULL,
  station = NULL,
  verbose = FALSE,
  timezone = "US/Alaska"
)

Arguments

chosen_date

Date or charater. Formatted as "YYYY-MM-DD"

latitude

Numeric. Fill in only if survey and station are not entered. latitude in either decimal degrees or a character latitude in degrees and decimal minutes

longitude

Numeric. Fill in only if survey and station are not entered. Longitude in either decimal degrees or a character longitude in degrees and decimal minutes

survey

Character. Fill in only if latitude and longitude are not entered. A character string of the survey you are interested in reivewing. Options are those from public_data$survey, which are "AI", "GOA", "EBS", "NBS", "BSS".

station

Character. Fill in only if latitude and longitude are not entered. A character string of the current station name (as a grid cell; e.g., "264-85")

verbose

Logical. Default = FALSE. If you would like a readout of what the file looks like in the console, set to TRUE.

timezone

Character. Default = "US/Alaska." Other options include: "US/Aleutian"

Value

Time of sunrise and sunset in text. Also shows a pop-up with sunrise and sunset times.

Examples

get_sunrise_sunset(chosen_date = Sys.Date(),
                   latitude = 63.3,
                   longitude = -170.5)
#> Using latitude and longitude to calcualte sunrise and sunset. 
#> Sunrise is at 7:37 AKST
#> Sunset is at 23:05 AKST
get_sunrise_sunset(chosen_date = as.character(Sys.Date()),
                   latitude = 63.3,
                   longitude = -170.5)
#> Using latitude and longitude to calcualte sunrise and sunset. 
#> Sunrise is at 7:37 AKST
#> Sunset is at 23:05 AKST
get_sunrise_sunset(chosen_date = Sys.Date(),
                   survey = "EBS",
                   station = "I-13")
#> Using average survey station location information (lat = 57.6709745126945, lon = -160.265447567568) to calcualte sunrise and sunset. 
#> Sunrise is at 7:18 AKST
#> Sunset is at 22:02 AKST
get_sunrise_sunset(chosen_date = Sys.Date(),
                   survey = "GOA",
                   station = "7-7")
#> Using average survey station location information (lat = 52.6006646808511, lon = -169.436484042553) to calcualte sunrise and sunset. 
#> Sunrise is at 8:09 AKST
#> Sunset is at 22:25 AKST
get_sunrise_sunset(chosen_date = "2023-06-10",
                   survey = "AI",
                   station = "324-73",
                   timezone = "US/Aleutian")
#> Using average survey station location information (lat = 54.2058070714286, lon = -165.745076285714) to calcualte sunrise and sunset. 
#> Sunrise is at 5:30 AKST
#> Sunset is at 22:35 AKST